Refactor info plugin and add tests
