+++ docs/providers/aws/cli-reference/info.md
@@ -69,11 +69,11 @@ functions:
   my-serverless-service-dev-hello: arn:aws:lambda:us-east-1:377024778620:function:my-serverless-service-dev-hello
 
 Stack Outputs
-CloudFrontUrl: d2d10e2tyk1pei.cloudfront.net
-ListScreenshotsLambdaFunctionArn: arn:aws:lambda:us-east-1:377024778620:function:lambda-screenshots-dev-listScreenshots
-ScreenshotBucket: dev-svdgraaf-screenshots
-CreateThumbnailsLambdaFunctionArn: arn:aws:lambda:us-east-1:377024778620:function:lambda-screenshots-dev-createThumbnails
-TakeScreenshotLambdaFunctionArn: arn:aws:lambda:us-east-1:377024778620:function:lambda-screenshots-dev-takeScreenshot
-ServiceEndpoint: https://12341jc801.execute-api.us-east-1.amazonaws.com/dev
-ServerlessDeploymentBucketName: lambda-screenshots-dev-serverlessdeploymentbucket-15b7pkc04f98a
+  CloudFrontUrl: d2d10e2tyk1pei.cloudfront.net
+  ListScreenshotsLambdaFunctionArn: arn:aws:lambda:us-east-1:377024778620:function:lambda-screenshots-dev-listScreenshots
+  ScreenshotBucket: dev-svdgraaf-screenshots
+  CreateThumbnailsLambdaFunctionArn: arn:aws:lambda:us-east-1:377024778620:function:lambda-screenshots-dev-createThumbnails
+  TakeScreenshotLambdaFunctionArn: arn:aws:lambda:us-east-1:377024778620:function:lambda-screenshots-dev-takeScreenshot
+  ServiceEndpoint: https://12341jc801.execute-api.us-east-1.amazonaws.com/dev
+  ServerlessDeploymentBucketName: lambda-screenshots-dev-serverlessdeploymentbucket-15b7pkc04f98a
 ```
